Extract out some proto files
