+++ nym-vpnd/src/command_interface/mod.rs
@@ -3,10 +3,11 @@
 
 mod config;
 mod connection_handler;
-mod credential_import;
 mod error;
 mod helpers;
 mod listener;
+mod proto_error;
+mod proto_response;
 mod socket_stream;
 mod start;
 mod status_broadcaster;
